optional.ts => helpers.ts, created check_user_token/scope_check functions, implemented and tested helper functions in albums.ts
